+++ flutter_reach_five/lib/helpers/provider_creator_converter.dart
@@ -2,7 +2,7 @@ import 'package:flutter_reach_five_platform_interface/flutter_reach_five_platfor
 
 import '../models/provider_creator.dart';
 
-/// Used to be the link between a [ProviderCreatorInterface]
+/// Link between a [ProviderCreatorInterface]
 /// and the [ProviderCreator] exported from this package
 class ProviderCreatorConverter {
   /// convert a [ProviderCreator] to a [ProviderCreatorInterface]